trace: Fix msvc compat

+++ libs/trace/src/sink_store.c
@@ -170,7 +170,7 @@ static void trace_sink_store_event_begin(
     evt->timeStart = time_steady_clock();
     evt->id        = trace_id_register(s, id);
     evt->color     = (u8)color;
-    evt->msgLength = math_min(msg.size, array_elems(evt->msgData));
+    evt->msgLength = (u8)math_min(msg.size, array_elems(evt->msgData));
     mem_cpy(array_mem(evt->msgData), mem_create(msg.ptr, evt->msgLength));
   }
   thread_spinlock_unlock(&evt->lock);
